AI_Voice_Assistant/requirements.txt
google-generativeai 
pyttsx3 
SpeechRecognition 
pyaudio